Refactor config code to be consistent in location
Change-type: patch
