+++ test/13-device-config.spec.coffee
@@ -9,7 +9,7 @@ prepare = require './lib/prepare'
 fsUtils = require '../src/lib/fs-utils'
 
 DeviceConfig = require '../src/device-config'
-{ ExtlinuxConfigBackend, RPiConfigBackend } = require '../src/lib/config-backend'
+{ ExtlinuxConfigBackend, RPiConfigBackend } = require '../src/config/backend'
 
 extlinuxBackend = new ExtlinuxConfigBackend()
 rpiConfigBackend = new RPiConfigBackend()
